# openapi.yaml
openapi: 3.0.3
info:
  title: Coachwise API
  version: "0.1.0"
  description: >
    Early-stage REST API for Coachwise. All non-auth routes require
    `Authorization: Bearer <access_token>`.
servers:
  - url: http://localhost:3000
paths:
...
  /users:
    get:
      summary: List users
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: username
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
  /users/me:
    get:
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      summary: Update current user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      summary: Delete current user
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Deleted
  /users/{id}:
    get:
      summary: Get user by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
  /feeds:
    get:
      summary: List feeds
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: offset
          schema:
            type: integer
        - in: query
          name: user_id
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Feeds
          content:
            application/json:
              schema:
                type: object
                properties:
                  feeds:
                    type: array
                    items:
                      $ref: '#/components/schemas/Feed'
                  total_count:
                    type: integer
    post:
      summary: Create feed post
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedCreateForm'
      responses:
        "200":
          description: Created feed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Feed'
  /feeds/{id}:
    get:
      summary: Get feed
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Feed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Feed'
    delete:
      summary: Delete feed
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Deleted
  /feeds/{id}/like:
    post:
      summary: Like feed
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Liked
    delete:
      summary: Unlike feed
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Unliked
  /feeds/{id}/comments:
    get:
      summary: List feed comments
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: object
                properties:
                  comments:
                    type: array
                    items:
                      $ref: '#/components/schemas/FeedComment'
    post:
      summary: Create feed comment
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedCommentForm'
      responses:
        "200":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedComment'
  /media:
    post:
      summary: Create media metadata
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaCreateForm'
      responses:
        "200":
          description: Created media
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Media'
    get:
      summary: List media for current user
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: offset
          schema:
            type: integer
      responses:
        "200":
          description: Media list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Media'
  /messages:
    post:
      summary: Send a message
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageCreateForm'
      responses:
        "200":
          description: Created message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
...
  /messages/{peer_id}:
    get:
      summary: List messages with a peer
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: peer_id
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: offset
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
  /messages/{peer_id}/read:
    post:
      summary: Mark thread as read
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: peer_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Success
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        job_title:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        first_name:
          type: string
          nullable: true
        last_name:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        avatar_id:
          type: string
          format: uuid
          nullable: true
        avatar:
          type: object
          properties:
            url:
              type: string
              nullable: true
            filename:
              type: string
              nullable: true
        status:
          type: string
        password_expired:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserUpdate:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        job_title:
          type: string
        bio:
          type: string
        phone:
          type: string
        username:
          type: string
...
    Tag:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        normalized_name:
          type: string
        is_system:
          type: boolean
        created_at:
          type: string
          format: date-time
    FeedMedia:
      type: object
      properties:
        id:
          type: string
          format: uuid
        feed_id:
          type: string
          format: uuid
        kind:
          type: string
          enum: [IMAGE, VIDEO]
        url:
          type: string
        thumbnail_url:
          type: string
          nullable: true
        order_index:
          type: integer
        created_at:
          type: string
          format: date-time
    Feed:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        body:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        visibility:
          type: string
          enum: [PUBLIC, PRIVATE]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        media:
          type: array
          items:
            $ref: '#/components/schemas/FeedMedia'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        like_count:
          type: integer
        comment_count:
          type: integer
        liked:
          type: boolean
    FeedCreateForm:
      type: object
      properties:
        body:
          type: string
        location:
          type: string
        visibility:
          type: string
          enum: [PUBLIC, PRIVATE]
        media:
          type: array
          items:
            $ref: '#/components/schemas/FeedMediaForm'
        tags:
          type: array
          items:
            type: string
    FeedMediaForm:
      type: object
      required: [kind, url]
      properties:
        kind:
          type: string
          enum: [IMAGE, VIDEO]
        url:
          type: string
        thumbnail_url:
          type: string
        order_index:
          type: integer
    FeedCommentForm:
      type: object
      required: [body]
      properties:
        body:
          type: string
    FeedComment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        feed_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        body:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sender_id:
          type: string
          format: uuid
        recipient_id:
          type: string
          format: uuid
        body:
          type: string
        media_id:
          type: string
          format: uuid
          nullable: true
        read_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    Media:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
        filename:
          type: string
        created_at:
          type: string
          format: date-time
    MediaCreateForm:
      type: object
      required: [url, filename]
      properties:
        url:
          type: string
        filename:
          type: string
    MessageCreateForm:
      type: object
      required: [recipient_id, body]
      properties:
        recipient_id:
          type: string
          format: uuid
        body:
          type: string
        media_id:
          type: string
          format: uuid
          nullable: true
...